SQL Queries
TripDiveView: full outer join of trips and dives. SQLite only supports
inner and left joins so this is a convenience view because not every
trip has a dive and some dives may fall outside of trips.

Boobies_GPS_TDR: work in progress. Trying to decide how to best merge
GPS and TDR data